+++ surepy/data/locdata.py
@@ -19,11 +19,10 @@ class LocData():
 
     Data consist of individual elements being either localizations or other LocData objects. Both, localizations and
     Locdata objects have properties. Properties come from the original data or are added by analysis procedures.
-    Analysis classes can take LocData objects as input on which to perform their action.
 
     Parameters
     ----------
-    references : LocData or list(LocData) or None
+    references : LocData, list(LocData), or None
         A locData reference or an array with references to locData objects referring to the selected localizations
         in dataset.
     dataframe : Pandas DataFrame or None
@@ -36,9 +35,9 @@ class LocData():
     Attributes
     ----------
     count : int
-        A counter for counting locdata instantiations (class attribute).
+        A counter for counting LocData instantiations (class attribute).
 
-    references : LocData or list(LocData) or None
+    references : LocData, list(LocData) or None
         A locData reference or an array with references to locData objects referring to the selected localizations
         in dataframe.
     dataframe : Pandas DataFrame or None
@@ -283,7 +282,6 @@ class LocData():
        dataframe. """
        if isinstance(self.references, LocData):
            df = self.references.data.iloc[self.indices]
-           df = df.reset_index()
            df = pd.merge(df, self.dataframe, left_index=True, right_index=True, how='outer')
            return df
        else:
@@ -318,16 +316,12 @@ class LocData():
         if self.references is None:
             return 0
         elif isinstance(self.references, LocData):
-            self.dataframe = self.references.data.iloc[self.indices]
-            self.dataframe = self.dataframe.reset_index()
+            self.dataframe = self.data
             self.indices = None
             self.references = None
             return 1
         else:
-            self.dataframe = self.dataframe.reset_index()
-            self.indices = None
-            self.references = None
-            return 1
+            raise ValueError('reference has undefined value.')
 
 
     def print_meta(self):
@@ -352,5 +346,3 @@ class LocData():
         meta_.frame_count = self.meta.frame_count
 
         print (text_format.MessageToString(meta_))
-
-